+++ src/Visualizer/vboindexer.cpp
@@ -9,7 +9,7 @@
 
 
 // Returns true iif v1 can be considered equal to v2
-bool is_near(float v1, float v2){
+bool is_near(double v1, double v2){
 	return fabs( v1-v2 ) < 0.01f;
 }
